chore: update tests for playwright
